update p_grad_c docstring to have an explanation of what it is, and remove the pgradc translate file, moving the TranslatePGradC class into TranslateDynCore
